requirements.txt
boto3
regex-engine
apache-airflow[amazon]
